% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varkitExtra-package.R
\docType{package}
\name{varkitExtra-package}
\alias{varkitExtra}
\alias{varkitExtra-package}
\title{varkitExtra: VarKit Extra: Additional features for VarKit}
\description{
To learn more about varkit, start with the vignette:
\code{browseVignettes(package = "varkit")}
}
\author{
\strong{Maintainer}: Dennie te Molder \email{dennie.temolder@wur.nl} (\href{https://orcid.org/0000-0003-2857-2812}{ORCID})

}
\keyword{internal}
